requirements.txt
joblib==1.4.2
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
torch==2.6.0
tqdm==4.67.1
